# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Papafi Backend API",
    "version": "1.0.0",
    "description": "Papafi wallet backend API for frontend and mobile clients. Documents implemented Express routes only."
  },
  "servers": [
    { "url": "https://api-staging.papifi.com" },
    { "url": "https://api.papifi.com" },
    { "url": "http://localhost:3000" }
  ],
...
  "paths": {
...
    "/api/transaction": {
      "get": {
        "tags": ["Transactions"],
        "summary": "List wallet and optional card transactions",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "walletId", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "cardId", "in": "query", "required": false, "schema": { "type": "string", "format": "uuid" } },
          { "name": "type", "in": "query", "required": false, "schema": { "type": "string", "enum": ["sent", "received"] } },
          { "name": "startDate", "in": "query", "required": false, "schema": { "type": "string", "format": "date-time" } },
          { "name": "endDate", "in": "query", "required": false, "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": { "description": "Transactions returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionListResponse" } } } },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
    "/api/transaction/log": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Internal wallet transfer",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/InternalTransferRequest" } } }
        },
        "responses": {
          "201": { "description": "Transfer created.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionDuplicateResponse" } } } },
          "200": { "description": "Duplicate transfer retry returned.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransactionDuplicateResponse" } } } },
          "400": { "$ref": "#/components/responses/OkFalseError" },
          "403": { "$ref": "#/components/responses/OkFalseError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Required for money movement endpoints. Must be a stable unique string of at least 8 characters per client operation. The backend also accepts idempotencyKey in JSON body.",
        "schema": { "type": "string", "minLength": 8 },
        "example": "wd_01hzz6n5qx9m8v2k"
      },
...
    },
    "responses": {
...
      "OkFalseError": { "description": "Error response.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkFalseError" }, "example": { "ok": false, "message": "Invalid transaction PIN" } } } },
...
    },
    "schemas": {
      "Currency": { "type": "string", "enum": ["NGN", "USD", "GBP"] },
...
      "OkFalseError": { "type": "object", "properties": { "ok": { "type": "boolean", "example": false }, "message": { "type": "string" } }, "required": ["ok", "message"] },
...
      "Transaction": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "userId": { "type": "string", "format": "uuid" }, "senderWalletId": { "type": "string", "format": "uuid", "nullable": true }, "recipientWalletId": { "type": "string", "format": "uuid", "nullable": true }, "type": { "type": "string", "enum": ["deposit", "withdrawal", "transfer"] }, "amount": { "type": "string", "example": "5000.00" }, "currency": { "$ref": "#/components/schemas/Currency" }, "status": { "type": "string" }, "reference": { "type": "string", "nullable": true }, "provider": { "type": "string", "nullable": true }, "providerReference": { "type": "string", "nullable": true }, "providerStatus": { "type": "string", "nullable": true }, "reconciliationStatus": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "InternalTransferRequest": { "type": "object", "required": ["senderWalletId", "recipientWalletId", "amount", "currency", "transactionPin"], "properties": { "senderWalletId": { "type": "string", "format": "uuid" }, "recipientWalletId": { "type": "string", "format": "uuid" }, "amount": { "type": "number", "example": 1000 }, "currency": { "$ref": "#/components/schemas/Currency" }, "description": { "type": "string" }, "transactionPin": { "type": "string", "example": "1234" }, "idempotencyKey": { "type": "string" } } },
      "TransactionDuplicateResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "transaction": { "$ref": "#/components/schemas/Transaction" }, "duplicate": { "type": "boolean" } } },
      "TransactionListResponse": { "type": "object", "properties": { "ok": { "type": "boolean" }, "count": { "type": "integer" }, "transactions": { "type": "array", "items": { "type": "object", "additionalProperties": true } } } },
...
    }
  }
}